+++ src/main/java/com/discount/Application.java
@@ -1,4 +1,4 @@
-package com.discount.java11;
+package com.discount;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
