+++ src/main/java/com/discount/Entity/Discount.java
@@ -1,0 +1,4 @@
+package com.discount.Entity;
+
+public class Discount {
+}
