+++ src/main/java/com/discount/Entity/Role.java
@@ -1,4 +1,4 @@
-package com.discount.java11.Entity;
+package com.discount.Entity;
 
 public enum Role {ADMIN, USER,ANALISIST
 }
