--- src/main/java/com/discount/java11/Entity/Discount.java
@@ -1,4 +1,0 @@
-package com.discount.java11.Entity;
-
-public class Discount {
-}
