--- src/main/java/com/discount/java11/Security/ConfigureUIServiceInitListener.java
@@ -1,33 +1,0 @@
-package com.discount.java11.Security;
-
-import com.vaadin.flow.component.UI;
-import com.vaadin.flow.router.BeforeEnterEvent;
-import com.vaadin.flow.server.ServiceInitEvent;
-import com.vaadin.flow.server.VaadinServiceInitListener;
-import org.springframework.stereotype.Component;
-import com.discount.java11.Security.ui.LoginView;
-
-@Component
-public class ConfigureUIServiceInitListener implements VaadinServiceInitListener {
-
-	@Override
-	public void serviceInit(ServiceInitEvent event) {
-		event.getSource().addUIInitListener(uiEvent -> {
-			final UI ui = uiEvent.getUI();
-			ui.addBeforeEnterListener(this::beforeEnter);
-		});
-	}
-
-	/**
-	 * Reroutes the user if (s)he is not authorized to access the view.
-	 *
-	 * @param event
-	 *            before navigation event with event details
-	 */
-	private void beforeEnter(BeforeEnterEvent event) {
-		if (!LoginView.class.equals(event.getNavigationTarget())
-		    && !SecurityUtils.isUserLoggedIn()) {
-			event.rerouteTo(LoginView.class);
-		}
-	}
-}
