--- src/main/java/com/discount/java11/Security/ui/MainView.java
@@ -1,21 +1,0 @@
-package com.discount.java11.Security.ui;
-
-import com.vaadin.flow.component.button.Button;
-import com.vaadin.flow.component.notification.Notification;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import com.vaadin.flow.component.orderedlayout.VerticalLayout;
-import com.vaadin.flow.router.Route;
-import com.vaadin.flow.server.PWA;
-
-@Route
-@PWA(name = "Project Base for Vaadin Flow with Spring", shortName = "Project Base")
-public class MainView extends VerticalLayout {
-
-    public MainView(@Autowired MessageBean bean) {
-        Button button = new Button("Click me",
-                e -> Notification.show(bean.getMessage()));
-        add(button);
-    }
-
-}
