--- src/main/java/com/discount/java11/Security/ui/MessageBean.java
@@ -1,12 +1,0 @@
-package com.discount.java11.Security.ui;
-
-import java.time.LocalTime;
-import org.springframework.stereotype.Service;
-
-@Service
-public class MessageBean {
-
-    public String getMessage() {
-        return "Button was clicked at " + LocalTime.now();
-    }
-}
