+++ src/test/java/com/discount/ApplicationTests.java
@@ -1,4 +1,4 @@
-package com.discount.java11;
+package com.discount;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
